+++ Changelog.md
@@ -4,6 +4,8 @@
 
 - Calendar versioning, `YY.MINOR.MICRO`
 
+- If the file system is activate, system activity will be logged
+
 - Build number based on commit hash
 
 Due to the huge time skip and architectural changes, I've (retroactively) switched to calendar versioning with ``v0.1`` now known as ``v20.1`` as well.